Memperbaiki bug pada Leaderboard dan juga menambahkan fitur agar register tidak bisa pada username yang sama
