+++ HighTrivia/src/pages/Register.js
@@ -1,17 +1,18 @@
 import React from "react";
 import FormLogin from "../components/FormLogin";
 import access from "../utils/access";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 
-function Register(props) {
+function Register() {
   const dispatch = useDispatch();
+  const users = useSelector(state => state.access.users);
 
   const onClickSubmit = event => {
     event.preventDefault();
     const username = event.target.username.value.trim();
     const password = event.target.password.value.trim();
     if (username !== "" && password !== "") {
-      access.register(username, password, dispatch);
+      access.register(username, password, users, dispatch);
     } else {
       alert("Tolong diisi dengan benar");
     }
